+++ libs/qwt/src/qwt_plot_spectrocurve.cpp
@@ -235,9 +235,9 @@ double QwtPlotSpectroCurve::penWidth() const
 
    \sa drawDots()
  */
-void QwtPlotSpectroCurve::drawSeries( QPainter* painter,
+void QwtPlotSpectroCurve::drawSeries(QPainter* painter,
     const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-    const QRectF& canvasRect, int from, int to ) const
+    const QRectF& canvasRect, size_t from, size_t to ) const
 {
     if ( !painter || dataSize() <= 0 )
         return;
