+++ libs/qwt/src/qwt_plot_vectorfield.h
@@ -136,7 +136,7 @@ class QWT_EXPORT QwtPlotVectorField
 
     virtual void drawSeries( QPainter*,
         const QwtScaleMap& xMap, const QwtScaleMap& yMap,
-        const QRectF& canvasRect, int from, int to ) const QWT_OVERRIDE;
+        const QRectF& canvasRect, size_t from, size_t to ) const QWT_OVERRIDE;
 
     virtual int rtti() const QWT_OVERRIDE;
