+++ libs/qwt/src/qwt_polar_layout.cpp
@@ -348,7 +348,7 @@ void QwtPolarLayout::activate( const QwtPolarPlot* plot,
 
     QRectF rect( boundingRect ); // undistributed rest of the plot rect
     rect.adjust( m_data->margin, m_data->margin,
-        -m_data->margin, -m_data->margin );
+        -((int)m_data->margin), -(int)(m_data->margin));
 
     // We extract all layout relevant data from the widgets
     // and save them to m_data->layoutData.
